@@ -1,7 +1,8 @@
 source 'https://rubygems.org'
-ruby '2.2.4'
+ruby '2.4.1'
 
-gem 'rails', '~> 4.2.7'
+gem 'rails', '~> 5.2.0'
+gem 'activerecord-postgis-adapter'
 gem 'pg'
 gem 'uglifier'
 gem 'sass-rails'
@@ -10,10 +11,9 @@ gem 'sass-rails'
 # Use jquery as the JavaScript library
 gem 'jquery-rails'
 gem 'jquery-ui-rails'
-gem 'puma'
-gem 'actionpack-action_caching'
 gem 'rails-jquery-autocomplete'
-gem 'rollbar', '~> 2.7.1'
+gem 'puma'
+gem 'rollbar'
 
 # Use Unicorn as the app server
 # gem 'unicorn'
@@ -22,12 +22,13 @@ gem 'rollbar', '~> 2.7.1'
 # gem 'capistrano-rails', group: :development
 
 group :development, :test do
+  gem 'pry'
   gem 'byebug'
   gem 'query_diet'
   gem 'bullet'
   # Access an IRB console on exception pages or by using <%= console %> in views
-  gem 'web-console', '~> 2.0'
-  gem 'rspec-rails', '~> 3.0'
+  gem 'web-console'
+  gem 'rspec-rails'
   gem 'capybara'
   gem 'active_record_query_trace'
   gem 'derailed'
